Adds Support For Putting Funcs on Prototype
Objects with functions on them are not deeply comparable. The previous
release introduced the ability to cast an immutable toObject with an
option to strip the functions away.

This adds support for putting functions on the instances prototype,
which makes them deeply comparable without needing to cast.
